Weaviate/JobFinder/requirements.txt
# Core dependencies for JobFinder project
fastapi
uvicorn
weaviate-client
pandas
requests
streamlit
pydantic
